requirements.txt
flask>=2.3.0
python-telegram-bot>=20.0